wip: add basic character search method
Add method to search for a character based on name and server.
Currently very dumb, and not a good way to do it, but proves the API works as expected.

Attached RealApiTests to use to hit the API, this should be commented out or removed later on, so it isn't run in normal tests.